starting on one strategy for keeping the dashboard live and up to date
One program on the server runs two programs in parallel. One will be a flask app to handle requests and read from the stored data, the other will be a scraper that constantly scrapes the web and saves that information for the flask app. The desktop client will ping the server every so often for live data updates.